# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StePanel API
  version: 0.6.0
  description: |
    Authenticated server operations, service inventory, and migration API.
    Mutating requests require the `stepanel_session` cookie and the matching
    `stepanel_csrf` value in the `X-CSRF-Token` header or `csrf` form field.
    Responses include an `X-Request-ID` correlation identifier.
servers:
  - url: http://127.0.0.1:8090
security:
  - sessionCookie: []
paths:
...
  /api/sites/resources/{site}:
    get:
      summary: Inspect a per-site resource profile and its enforcement scope
      parameters: [{name: site, in: path, required: true, schema: {type: string}}]
      responses: {'200': {description: Desired/applied resource profile}}
    put:
      summary: Apply CPU, memory, I/O, task and PHP-FPM worker ceilings for a site
      parameters: [{name: site, in: path, required: true, schema: {type: string}}]
      responses: {'202': {description: Systemd slice and FPM profile applied}}
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: stepanel_session